renamed files and fixed filter function
